Split out batch, tool and util functions; also allow wrapper generation directly from the reanalyse and install commands for a speedup
