+++ aclimatise_automation/util.py
@@ -11,14 +11,16 @@ from contextlib import contextmanager, redirect_stderr, redirect_stdout
 from itertools import chain
 from logging import getLogger
 from multiprocessing import Lock
-from typing import List
+from typing import List, Collection
 
 from packaging.version import parse
 
 import requests
-from aclimatise import explore_command
+from aclimatise import explore_command, Command, WrapperGenerator
 from aclimatise.converter.yml import YmlGenerator
 from aclimatise.execution.docker import DockerExecutor
+
+from aclimatise_automation.metadata import BaseCampMeta
 from aclimatise_automation.yml import yaml
 from docker.models.containers import Container
 from git import Repo
@@ -130,6 +132,7 @@ def aclimatise_exe(
     container: Container,
     exe: str,
     out_dir: pathlib.Path,
+    wrapper_root: pathlib.Path  = None
 ):
     """
     Given an executable path, aclimatises it, and dumps the results in out_dir
@@ -144,7 +147,75 @@ def aclimatise_exe(
         # Rather than writing out the whole tree, which has redundant information, we instead take the top level command
         # which contains the entire tree, and serialize that
         gen.save_to_file(cmd, path)
+
+        if wrapper_root:
+            wrapper_from_command(
+                cmd=cmd,
+                command_path=path,
+                command_root=out_dir,
+                wrapper_root=wrapper_root
+            )
     except Exception as e:
         handle_exception()
 
+
     logger.info("Successfully written to YAML".format(exe))
+
+def calculate_metadata(
+        test=False,
+        filter_r=False,
+        filter_type: Collection[str] = {"CommandLineTool"},
+) -> BaseCampMeta:
+    """
+    Generates a new metadata file, which is basically a specification for an automation run
+    :param test: If we're in test mode
+    :param filter_r: If true, filter out R and bioconductor packages
+    :param filter_type: A list of toolClasses strings to select, or "none" to disable filtering
+    :return:
+    """
+
+    # The package names are keys to the output dict
+    if test:
+        images = ["bwa=0.7.17"]
+    else:
+        images = latest_biocontainers(filter_r=filter_r, filter_type=filter_type)
+
+    return BaseCampMeta(
+        packages=images, aclimatise_version=latest_package_version("aclimatise")
+    )
+
+
+def wrapper_from_command(cmd: Command, command_path: pathlib.Path, command_root:pathlib.Path, wrapper_root : pathlib.Path):
+    """
+    Given an already generated command, dump the wrappers
+    :param cmd:
+    :param command_path:
+    :param command_root:
+    :param wrapper_root:
+    :return:
+    """
+    output_path = pathlib.Path(wrapper_root) / command_path.parent.relative_to(command_root)
+
+    output_path.mkdir(parents=True, exist_ok=True)
+
+    try:
+        generators = [Gen() for Gen in WrapperGenerator.__subclasses__()]
+        for cmd in cmd.command_tree():
+            if len(cmd.subcommands) > 0:
+                # Since we're dumping directly usable tool definitions, it doesn't make sense to dump the parent
+                # commands like "samtools" rather than "samtools index", so skip them
+                continue
+
+            # Also, if we are dumping, we disconnect each Command from the command tree to simplify the output
+            cmd.parent = None
+            cmd.subcommands = []
+
+            for gen in generators:
+                path = executable_dir / (cmd.as_filename + gen.suffix)
+                gen.save_to_file(cmd, path)
+                logger.info(
+                    "{} converted to {}".format(" ".join(cmd.command), gen.suffix)
+                )
+    except Exception as e:
+        logger.error(handle_exception())
+
